requirements.txt
os
requests
azure.storage.blob
time
json
io